Simplified Chinese localization  (#193)
* make files

Ready Files

* Update FormBoxelSearch.zh-Hans.resx

* Update FormGroundTarget.zh-Hans.resx

* Update FormSetKeyChord.zh-Hans.resx

* invalid?

Is this an invalid multilingual file?

* Update PlotBodyInfo.zh-Hans.resx

* Update PlotBodyInfo.zh-Hans.resx

* Update PlotBioSystem.zh-Hans.resx

* Update PlotBioStatus.zh-Hans.resx

* Update FormSphereLimit.zh-Hans.resx

* Update PlotHumanSite.zh-Hans.resx

* Update PlotGrounded.zh-Hans.resx

* Update PlotFSSInfo.zh-Hans.resx

* Update PlotFSS.zh-Hans.resx

* Update PlotFlightWarning.zh-Hans.resx

* Update Codex.zh-Hans.resx

Attempt to translate proper nouns, taking into account the hints from the comments. Temporarily retain the original text, use "()" to indicate paraphrasing, and use "{}" to indicate transliteration.
尝试翻译专有名词，结合comment的提示，暂时保留原文，使用“()”表示意译，使用“{}”表示音译。

* Update PlotJumpInfo.zh-Hans.resx

* Update PlotPriorScans.zh-Hans.resx

* Update PlotSphericalSearch.zh-Hans.resx

* Update PlotTrackers.zh-Hans.resx

* Update CodexShort.zh-Hans.resx

Because it is necessary to retain the 3-letter code, it cannot be translated.
因要保留3字母，无法翻译

* Update Economies.zh-Hans.resx

Attempt to replace the translation without retaining the original text.
尝试不保留原文替换翻译

* Update FactionStates.zh-Hans.resx

Attempt to replace the translation without retaining the original text.
尝试不保留原文来替换翻译
Thargoid 萨格族

* Update KeyChords.zh-Hans.resx

* Update Reputation.zh-Hans.resx

* Update Misc.zh-Hans.resx

The data starting from "FormBoxelSearch_Disable" afterward, which window does it belong to? It seems to be duplicated with "\forms\FormBoxelSearch.resx"?

* fix errors.

Fix code errors.
3 typos.

* updata

Simplified Chinese localization

* Fix code errors caused by version confusion

The code structure and version differences were messed up by "GitHub Desktop," or it might be because I'm not very skilled at using this tool.
In the end, I manually replaced and fixed them one by one.

* Update PlotFSSInfo.zh-Hans.resx

Fix a minor error.

* fix error

Delete duplicate added entries

* Add hotkeys.

Add according to the existing hotkeys in the original English language

* Update PlotFSSInfo.zh-Hans.resx

Improve the style.

* updata newfiles

Translate the files added after merging and optimize the previous content

* Update Misc.zh-Hans.resx

* Update PlotFSS.zh-Hans.resx

text polishing

* text polishing